refactor: camera implementation in one file
